ref: split users&rooms into admin-settings and basic users/rooms
