+++ tests/interactions/dashboard-rooms.ts
@@ -1,49 +1,10 @@
 import { expect, Page } from '@playwright/test';
 import { TEST_IDS } from '../../src/test-ids';
-
-import * as types from '../support/types';
-import * as formInteractions from './forms';
 import * as navigation from './navigation';
-import * as settingsInteractions from './settings';
-import { TIMEOUTS } from '../support/constants';
 
-export const create = async (page: Page, room: types.RoomData) => {
-  await navigation.goToRoomsSettings(page);
-
-  await formInteractions.clickButton(page, 'add-rooms-button');
-  await page.waitForSelector('input[name="room_name"]', { state: 'visible', timeout: TIMEOUTS.HALF_SECOND });
-  await formInteractions.fillForm(page, 'room-name', room.name);
-  await formInteractions.fillMarkdownForm(page, 'description_public', room.description);
-
-  const UserSelector = page.getByTestId('users-field');
-  await expect(UserSelector).toBeVisible();
-
-  for (const u of room.users) {
-    await UserSelector.locator('.MuiAutocomplete-popupIndicator').click();
-    await page.getByTestId(`select-option-${u.username}`).filter({ visible: true }).click();
-    await page.getByTestId(`select-option-${u.username}`).waitFor({ state: 'hidden' });
-  }
-
-  await formInteractions.clickButton(page, 'room-form-submit-button');
-  // Wait for form close AND server confirmation (form closes optimistically before API responds)
-  await page.waitForSelector('input[name="room_name"]', { state: 'hidden' });
-  await page.waitForLoadState('networkidle');
-
-  await expect(page.getByTestId('add-rooms-button')).toBeVisible();
-  await settingsInteractions.filter(page, { option: 'room_name', value: room.name });
-  await settingsInteractions.clearFilter(page);
-};
-
-export const remove = async (page: Page, room: types.RoomData) => {
-  await navigation.goToRoomsSettings(page);
-
-  await settingsInteractions.remove({ page, scope: 'rooms', filters: { option: 'room_name', value: room.name } });
-};
-
-// Search and Sort interactions for Rooms View (Home Page)
+// Search and Sort interactions for Rooms View (Home Page / Dashboard)
 
 export const openSearch = async (page: Page) => {
-  await navigation.goToHome(page);
   await page.getByTestId(TEST_IDS.SEARCH_BUTTON).filter({ visible: true }).click();
   const searchField = page.getByTestId(TEST_IDS.SEARCH_FIELD).locator('input');
   await expect(searchField).toBeVisible();
@@ -65,6 +26,8 @@ export const searchRooms = async (page: Page, query: string) => {
   }
 
   await searchField.fill(query);
+  // Wait for the rooms list to re-fetch after filling the search filter.
+  await page.waitForLoadState('networkidle');
 };
 
 export const clearSearch = async (page: Page) => {
@@ -74,9 +37,8 @@ export const clearSearch = async (page: Page) => {
   await page.waitForLoadState('networkidle');
 };
 
-export const openSort = async (page: Page) => {
-  await navigation.goToHome(page);
-
+// this function only opens the sort menu, but doesn't click on the dropdown
+export const openSortMenu = async (page: Page) => {
   const sortSelect = page.getByTestId(TEST_IDS.SORT_SELECT);
   if (!(await sortSelect.isVisible())) {
     await page.getByTestId(TEST_IDS.SORT_BUTTON).filter({ visible: true }).click();
@@ -84,12 +46,13 @@ export const openSort = async (page: Page) => {
   }
 };
 
-export const closeSort = async (page: Page) => {
+export const closeSortMenu = async (page: Page) => {
   await page.getByTestId(TEST_IDS.SORT_BUTTON).filter({ visible: true }).click();
   const sortSelect = page.getByTestId(TEST_IDS.SORT_SELECT);
   await expect(sortSelect).not.toBeVisible();
 };
 
+// this function also opens the dropdown
 export const selectSortOption = async (page: Page, sortValue: string) => {
   await page.getByTestId(TEST_IDS.SORT_SELECT).filter({ visible: true }).click();
   await page.getByTestId(`sort-option-${sortValue}`).filter({ visible: true }).click();
@@ -98,23 +61,20 @@ export const selectSortOption = async (page: Page, sortValue: string) => {
 export const toggleSortDirection = async (page: Page) =>
   await page.getByTestId(TEST_IDS.SORT_DIRECTION_BUTTON).filter({ visible: true }).click();
 
-export const getRoomCount = async (page: Page): Promise<number> => {
-  await navigation.goToHome(page);
-
+export const getCountOfRoomsDisplayed = async (page: Page): Promise<number> => {
   const roomCards = page.getByTestId(TEST_IDS.ROOM_CARD);
   // Rooms load asynchronously after the heading appears — poll until the count stabilizes.
   let lastCount = -1;
-  for (let i = 0; i < 10; i++) {
+  for (let i = 0; i < 5; i++) {
     const count = await roomCards.count();
     if (count === lastCount) break;
     lastCount = count;
-    await page.waitForTimeout(150);
+    await page.waitForTimeout(i * 50);
   }
   return lastCount < 0 ? 0 : lastCount;
 };
 
 export const getFirstRoomName = async (page: Page): Promise<string | null> => {
-  await navigation.goToHome(page);
   const namesOfRooms = page.getByTestId(TEST_IDS.ROOM_CARD).filter({ visible: true }).locator(page.locator('h3'));
   return await namesOfRooms.first().textContent();
 };
